Inline ActionBarModel into CharacterModel
