+++ common/src/main/java/com/wynntils/models/character/actionbar/event/CenterSegmentClearedEvent.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2022.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.wynn.model.actionbar.event;
+package com.wynntils.models.character.actionbar.event;
 
 import net.minecraftforge.eventbus.api.Event;
